Radon

RadonKA.radon — Function

radon(I, θs, μ=nothing)

Calculates the parallel Radon transform of the AbstractArray I. The first two dimensions are y and x. The third dimension is z, the rotational axis. size(I, 1) and size(I, 2) have to be equal and a even number. The Radon transform is rotated around the pixel size(I, 1) ÷ 2 + 1, so there is always a real center pixel! Works either with a AbstractArray{T, 3} or AbstractArray{T, 2}.

θs is a vector or range storing the angles in radians.

Exponential IRadon Transform

If μ != nothing, then the rays are attenuated with exp(-μ * dist) where dist is the distance to the circular boundary of the field of view. μ is in units of pixel length. So μ=1 corresponds to an attenuation of exp(-1) if propagated through one pixel.

In principle, all backends of KernelAbstractions.jl should work but are not tested. CUDA and CPU arrays are actively tested.

Please note: the implementation is not quite optimized for cache efficiency and it is a very naive algorithm. But still, it is fast!

See also iradon.

Example

The reason the sinogram has the value 1.41421 for the diagonal ray π/4 is, that such a diagonal travels a longer distance through the pixel.

julia> arr = zeros((4,4)); arr[3,3] = 1;

julia> radon(arr, [0, π/4, π/2])
3×3 view(::Array{Float64, 3}, :, :, 1) with eltype Float64:
 0.0  0.0      0.0
 1.0  1.41421  1.0
 0.0  0.0      0.0

IRadon

RadonKA.iradon — Function

iradon(sinogram, θs, μ=nothing)

Calculates the parallel inverse Radon transform of the sinogram. The first two dimensions are y and x. The third dimension is z, the rotational axis. Works either with a AbstractArray{T, 3} or AbstractArray{T, 2}. size(sinogram, 1) has to be an odd number. And size(sinogram, 2) has to be equal to length(angles). The inverse Radon transform is rotated around the pixel size(sinogram, 1) ÷ 2, so there is always a real center pixel!

θs is a vector or range storing the angles in radians.

Exponential IRadon Transform

If μ != nothing, then the rays are attenuated with exp(-μ * dist) where dist is the distance to the circular boundary of the field of view. μ is in units of pixel length. So μ=1 corresponds to an attenuation of exp(-1) if propagated through one pixel.

In principle, all backends of KernelAbstractions.jl should work but are not tested. CUDA and CPU arrays are actively tested.

See also radon.

Examples

julia> arr = zeros((5,2)); arr[2,:] .= 1; arr[4, :] .= 1
2-element view(::Matrix{Float64}, 4, :) with eltype Float64:
 1.0
 1.0

julia> iradon(arr, [0, π/2])
6×6 view(::Array{Float64, 3}, :, :, 1) with eltype Float64:
 0.0  0.0  0.0        0.0  0.0        0.0
 0.0  0.0  0.1        0.0  0.1        0.0
 0.0  0.1  0.2        0.1  0.2        0.0232051
 0.0  0.0  0.1        0.0  0.1        0.0
 0.0  0.1  0.2        0.1  0.2        0.0232051
 0.0  0.0  0.0232051  0.0  0.0232051  0.0

julia> iradon(arr, [0, π/2], 1) # exponential
6×6 view(::Array{Float64, 3}, :, :, 1) with eltype Float64:
 0.0  0.0         0.0         0.0        0.0         0.0
 0.0  0.0         0.00145226  0.0        0.00145226  0.0
 0.0  0.00145226  0.00789529  0.0107308  0.033117    0.0183994
 0.0  0.0         0.0107308   0.0        0.0107308   0.0
 0.0  0.00145226  0.033117    0.0107308  0.0583388   0.0183994
 0.0  0.0         0.0183994   0.0        0.0183994   0.0

RadonKA.filtered_backprojection — Function

filtered_backprojection(sinogram, θs)

Calculates the simple Filtered Backprojection in CT with applying a ramp filter in Fourier space.